Sistemata creazione, modifica e cancellazione: ora funzionano anche le categorie e i biglietti.

+++ src/main/java/it/unife/ingsw202324/EventManager/repositories/CategoriesRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CategoriesRepository extends JpaRepository<Categories, Long> {
+public interface CategoriesRepository extends JpaRepository<Categories, Integer> {
     List<Categories> findByDeleted(boolean deleted);
 }
